Lesson 37 - DeleteItemPermission

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -43,7 +43,17 @@ const Mutations = {
 
 	async deleteItem(parent, args, ctx, info) {
 		const where = { id: args.id };
-		const item = await ctx.db.query.item({ where }, `{id,title}`);
+		const item = await ctx.db.query.item({ where }, `{id title user{id}}`);
+
+		const ownsItem = item.user.id === ctx.request.userId;
+		const hasPermissions = ctx.request.user.permissions.some((permisssion) =>
+			[ 'ADMIN', 'ITEMDELETE' ].includes(permisssion)
+		);
+
+		if (!ownsItem || !hasPermission) {
+			throw new Error("You don't have enough permissions to do that");
+		}
+
 		return ctx.db.mutation.deleteItem({ where }, info);
 	},
